Move out all series-building related code from chartOptionsBuilder
JIRA: RAIL-1815
