+++ libs/sdk-ui-charts/src/highcharts/_to_refactor/chartOptionsBuilder.ts
@@ -1,7 +1,6 @@
 // (C) 2007-2020 GoodData Corporation
 import { colors2Object, numberFormat } from "@gooddata/numberjs";
 import {
-    DataValue,
     IAttributeDescriptor,
     IDataView,
     IMeasureDescriptor,
@@ -12,8 +11,8 @@ import cx from "classnames";
 import invariant from "ts-invariant";
 
 import {
-    DataViewFacade,
     BucketNames,
+    DataViewFacade,
     getDrillIntersection,
     IHeaderPredicate,
     IMappingHeader,
@@ -33,13 +32,14 @@ import { HEATMAP_DATA_POINTS_LIMIT, PIE_CHART_LIMIT } from "./constants/limits";
 import { findAttributeInDimension, findMeasureGroupInDimensions } from "./utils/executionResultHelper";
 import { IUnwrappedAttributeHeadersWithItems } from "./utils/types";
 
-import { IColorStrategy, getLighterColor } from "@gooddata/sdk-ui-vis-commons";
+import { IColorStrategy } from "@gooddata/sdk-ui-vis-commons";
 
 import {
     customEscape,
     isAreaChart,
     isBarChart,
     isBubbleChart,
+    isBulletChart,
     isChartSupported,
     isComboChart,
     isCssMultiLineTruncationSupported,
@@ -50,7 +50,6 @@ import {
     parseValue,
     stringifyChartTypes,
     unwrap,
-    isBulletChart,
 } from "./utils/common";
 import { setMeasuresToSecondaryAxis } from "./utils/dualAxis";
 import {
@@ -78,10 +77,7 @@ import compact from "lodash/compact";
 import get from "lodash/get";
 import includes from "lodash/includes";
 import isEmpty from "lodash/isEmpty";
-import isEqual from "lodash/isEqual";
-import isNil from "lodash/isNil";
 import isUndefined from "lodash/isUndefined";
-import last from "lodash/last";
 import range from "lodash/range";
 import without from "lodash/without";
 import omit from "lodash/omit";
@@ -90,14 +86,16 @@ import {
     IAxis,
     ICategory,
     IChartOptions,
-    IPatternObject,
     IPointData,
     ISeriesDataItem,
     ISeriesItem,
     ISeriesItemConfig,
 } from "../typings/unsafe";
-import { getBulletChartSeries } from "../chartTypes/bulletChart/bulletChartOptions";
-import { GRAY, WHITE, TRANSPARENT } from "./utils/color";
+import { getBulletChartSeries } from "../chartTypes/bulletChart/bulletChartSeries";
+import { getBubbleChartSeries } from "../chartTypes/bubbleChart/bubbleChartSeries";
+import { getTreemapStackedSeries } from "../chartTypes/treemap/treemapChartSeries";
+import { getHeatmapSeries } from "../chartTypes/heatmap/heatmapChartSeries";
+import { getScatterPlotSeries } from "../chartTypes/scatterPlot/scatterPlotSeries";
 
 const TOOLTIP_PADDING = 10;
 
@@ -136,19 +134,6 @@ const unsupportedStackingTypes = [
     VisualizationTypes.SCATTER,
     VisualizationTypes.BUBBLE,
 ];
-
-const nullColor: IPatternObject = {
-    pattern: {
-        path: {
-            d: "M 10 0 L 0 10 M 9 11 L 11 9 M 4 11 L 11 4 M -1 1 L 1 -1 M -1 6 L 6 -1",
-            stroke: GRAY,
-            strokeWidth: 1,
-            fill: WHITE,
-        },
-        width: 10,
-        height: 10,
-    },
-};
 
 export interface IValidationResult {
     dataTooLarge: boolean;
@@ -297,314 +282,6 @@ export function getSeriesItemData(
 
         return pointData;
     });
-}
-
-// eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
-export function getHeatmapSeries(
-    dv: DataViewFacade,
-    measureGroup: IMeasureGroupDescriptor["measureGroupHeader"],
-) {
-    const data: IPointData[] = [];
-    dv.rawData()
-        .twoDimData()
-        .forEach((rowItem: DataValue[], rowItemIndex: number) => {
-            rowItem.forEach((columnItem: DataValue, columnItemIndex: number) => {
-                const value: number = parseValue(String(columnItem));
-                const pointData: IPointData = { x: columnItemIndex, y: rowItemIndex, value };
-                if (isNil(value)) {
-                    data.push({
-                        ...pointData,
-                        borderWidth: 1,
-                        borderColor: GRAY,
-                        color: TRANSPARENT,
-                    });
-                    data.push({
-                        ...pointData,
-                        borderWidth: 0,
-                        pointPadding: 2,
-                        color: nullColor,
-                        // ignoredInDrillEventContext flag is used internally, not related to Highchart
-                        // to check and remove this null-value point in drill message
-                        ignoredInDrillEventContext: true,
-                    });
-                } else {
-                    data.push(pointData);
-                }
-            });
-        });
-
-    return [
-        {
-            name: measureGroup.items[0].measureHeaderItem.name,
-            data,
-            turboThreshold: 0,
-            yAxis: 0,
-            dataLabels: {
-                formatGD: unwrap(measureGroup.items[0]).format,
-            },
-            legendIndex: 0,
-        },
-    ];
-}
-
-export function getScatterPlotSeries(
-    dv: DataViewFacade,
-    // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
-    stackByAttribute: any,
-    colorStrategy: IColorStrategy,
-): any[] {
-    const primaryMeasuresBucketEmpty = dv.def().isBucketEmpty(BucketNames.MEASURES);
-    const secondaryMeasuresBucketEmpty = dv.def().isBucketEmpty(BucketNames.SECONDARY_MEASURES);
-
-    const data: ISeriesDataItem[] = dv
-        .rawData()
-        .twoDimData()
-        .map((seriesItem: string[], seriesIndex: number) => {
-            const values = seriesItem.map((value: string) => {
-                return parseValue(value);
-            });
-
-            return {
-                x: !primaryMeasuresBucketEmpty ? values[0] : 0,
-                y: !secondaryMeasuresBucketEmpty ? (primaryMeasuresBucketEmpty ? values[0] : values[1]) : 0,
-                name: stackByAttribute ? stackByAttribute.items[seriesIndex].attributeHeaderItem.name : "",
-            };
-        });
-
-    return [
-        {
-            turboThreshold: 0,
-            color: colorStrategy.getColorByIndex(0),
-            legendIndex: 0,
-            data,
-        },
-    ];
-}
-
-function getCountOfEmptyBuckets(bucketEmptyFlags: boolean[] = []) {
-    return bucketEmptyFlags.filter((bucketEmpyFlag) => bucketEmpyFlag).length;
-}
-
-export function getBubbleChartSeries(
-    dv: DataViewFacade,
-    measureGroup: IMeasureGroupDescriptor["measureGroupHeader"],
-    // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
-    stackByAttribute: any,
-    colorStrategy: IColorStrategy,
-): any[] {
-    const primaryMeasuresBucketEmpty = dv.def().isBucketEmpty(BucketNames.MEASURES);
-    const secondaryMeasuresBucketEmpty = dv.def().isBucketEmpty(BucketNames.SECONDARY_MEASURES);
-
-    return dv
-        .rawData()
-        .twoDimData()
-        .map((resData: any, index: number) => {
-            let data: any = [];
-            if (resData[0] !== null && resData[1] !== null && resData[2] !== null) {
-                const emptyBucketsCount = getCountOfEmptyBuckets([
-                    primaryMeasuresBucketEmpty,
-                    secondaryMeasuresBucketEmpty,
-                ]);
-                data = [
-                    {
-                        x: !primaryMeasuresBucketEmpty ? parseValue(resData[0]) : 0,
-                        y: !secondaryMeasuresBucketEmpty ? parseValue(resData[1 - emptyBucketsCount]) : 0,
-                        // we want to allow NaN on z to be able show bubble of default size when Size bucket is empty
-                        z: parseFloat(resData[2 - emptyBucketsCount]),
-                        format: unwrap(last(measureGroup.items)).format, // only for dataLabel format
-                    },
-                ];
-            }
-            return {
-                name: stackByAttribute ? stackByAttribute.items[index].attributeHeaderItem.name : "",
-                color: colorStrategy.getColorByIndex(index),
-                legendIndex: index,
-                data,
-            };
-        });
-}
-
-function getColorStep(valuesCount: number): number {
-    const MAX_COLOR_BRIGHTNESS = 0.8;
-    return MAX_COLOR_BRIGHTNESS / valuesCount;
-}
-
-function gradientPreviousGroup(solidColorLeafs: any[]): any[] {
-    const colorChange = getColorStep(solidColorLeafs.length);
-    return solidColorLeafs.map((leaf: any, index: number) => ({
-        ...leaf,
-        color: getLighterColor(leaf.color, colorChange * index),
-    }));
-}
-
-function getRootPoint(rootName: string, index: number, format: string, colorStrategy: IColorStrategy) {
-    return {
-        id: `${index}`,
-        name: rootName,
-        color: colorStrategy.getColorByIndex(index),
-        showInLegend: true,
-        legendIndex: index,
-        format,
-    };
-}
-
-function getLeafPoint(
-    stackByAttribute: any,
-    parentIndex: number,
-    seriesIndex: number,
-    data: any,
-    format: string,
-    colorStrategy: IColorStrategy,
-) {
-    return {
-        name: stackByAttribute.items[seriesIndex].attributeHeaderItem.name,
-        parent: `${parentIndex}`,
-        value: parseValue(data),
-        x: seriesIndex,
-        y: seriesIndex,
-        showInLegend: false,
-        color: colorStrategy.getColorByIndex(parentIndex),
-        format,
-    };
-}
-
-function isLastSerie(seriesIndex: number, dataLength: number) {
-    return seriesIndex === dataLength - 1;
-}
-
-export function getTreemapStackedSeriesDataWithViewBy(
-    dv: DataViewFacade,
-    measureGroup: IMeasureGroupDescriptor["measureGroupHeader"],
-    viewByAttribute: IUnwrappedAttributeHeadersWithItems,
-    stackByAttribute: IUnwrappedAttributeHeadersWithItems,
-    colorStrategy: IColorStrategy,
-): any[] {
-    const roots: any = [];
-    const leafs: any = [];
-    let rootId = -1;
-    let uncoloredLeafs: any = [];
-    let lastRoot: IResultAttributeHeader["attributeHeaderItem"] = null;
-
-    const executionResultData = dv.rawData().twoDimData();
-    const dataLength = executionResultData.length;
-    const format = unwrap(measureGroup.items[0]).format; // this configuration has only one measure
-
-    executionResultData.forEach((seriesItems: string[], seriesIndex: number) => {
-        const currentRoot = viewByAttribute.items[seriesIndex].attributeHeaderItem;
-
-        if (!isEqual(currentRoot, lastRoot)) {
-            // store previous group leafs
-            leafs.push(...gradientPreviousGroup(uncoloredLeafs));
-            rootId++;
-            lastRoot = currentRoot;
-            uncoloredLeafs = [];
-            // create parent for pasted leafs
-            const lastRootName: string = get(lastRoot, "name");
-            roots.push(getRootPoint(lastRootName, rootId, format, colorStrategy));
-        }
-        // create leafs which will be colored at the end of group
-        uncoloredLeafs.push(
-            getLeafPoint(stackByAttribute, rootId, seriesIndex, seriesItems[0], format, colorStrategy),
-        );
-
-        if (isLastSerie(seriesIndex, dataLength)) {
-            // store last group leafs
-            leafs.push(...gradientPreviousGroup(uncoloredLeafs));
-        }
-    });
-
-    return [...roots, ...leafs]; // roots need to be first items in data to keep legend working
-}
-
-export function getTreemapStackedSeriesDataWithMeasures(
-    dv: DataViewFacade,
-    measureGroup: IMeasureGroupDescriptor["measureGroupHeader"],
-    // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
-    stackByAttribute: any,
-    colorStrategy: IColorStrategy,
-): any[] {
-    let data: any = [];
-
-    measureGroup.items.reduce((data: any[], measureGroupItem: any, index: number) => {
-        data.push({
-            id: `${index}`,
-            name: measureGroupItem.measureHeaderItem.name,
-            format: measureGroupItem.measureHeaderItem.format,
-            color: colorStrategy.getColorByIndex(index),
-            showInLegend: true,
-            legendIndex: index,
-        });
-        return data;
-    }, data);
-
-    dv.rawData()
-        .twoDimData()
-        .forEach((seriesItems: string[], seriesIndex: number) => {
-            const colorChange = getColorStep(seriesItems.length);
-            const unsortedLeafs: any[] = [];
-            seriesItems.forEach((seriesItem: string, seriesItemIndex: number) => {
-                unsortedLeafs.push({
-                    name: stackByAttribute.items[seriesItemIndex].attributeHeaderItem.name,
-                    parent: `${seriesIndex}`,
-                    format: unwrap(measureGroup.items[seriesIndex]).format,
-                    value: parseValue(seriesItem),
-                    x: seriesIndex,
-                    y: seriesItemIndex,
-                    showInLegend: false,
-                });
-            });
-            const sortedLeafs = unsortedLeafs.sort((a: IPointData, b: IPointData) => b.value - a.value);
-
-            data = [
-                ...data,
-                ...sortedLeafs.map((leaf: IPointData, seriesItemIndex: number) => ({
-                    ...leaf,
-                    color: getLighterColor(
-                        colorStrategy.getColorByIndex(seriesIndex),
-                        colorChange * seriesItemIndex,
-                    ),
-                })),
-            ];
-        });
-
-    return data;
-}
-
-export function getTreemapStackedSeries(
-    dv: DataViewFacade,
-    measureGroup: IMeasureGroupDescriptor["measureGroupHeader"],
-    viewByAttribute: IUnwrappedAttributeHeadersWithItems,
-    stackByAttribute: IUnwrappedAttributeHeadersWithItems,
-    colorStrategy: IColorStrategy,
-): any[] {
-    let data = [];
-    if (viewByAttribute) {
-        data = getTreemapStackedSeriesDataWithViewBy(
-            dv,
-            measureGroup,
-            viewByAttribute,
-            stackByAttribute,
-            colorStrategy,
-        );
-    } else {
-        data = getTreemapStackedSeriesDataWithMeasures(dv, measureGroup, stackByAttribute, colorStrategy);
-    }
-
-    const seriesName = measureGroup.items
-        .map((wrappedMeasure: IMeasureDescriptor) => {
-            return unwrap(wrappedMeasure).name;
-        })
-        .join(", ");
-
-    return [
-        {
-            name: seriesName,
-            legendType: "point",
-            showInLegend: true,
-            data,
-            turboThreshold: 0,
-        },
-    ];
 }
 
 export function getSeries(
