+++ libs/sdk-ui-charts/src/highcharts/chartTypes/bulletChart/bulletChartSeries.ts
